Handle long profile step names better
Long profile step names would collide with the timing information
making both unreadable.

For step names that expand, any text that would overflow is
replaced with an ellipsis and a hover text is added with
the full step name.

For step names that don't expand, the text is forced to wrap.

These changes are particularly useful for SQL queries with
text that doesn't wrap, and the AspectJ profiling when
step names include full package, class, and method names.